refactor(api): clean up blog step from cars workflow
